pyproject.toml
[project]
name = "cc-proxy"
version = "0.1.0"
description = "MCP Server bridging Hermes Agent with local Claude Code sessions"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
cc-proxy = "cc_proxy.server:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
